feat(wiki): Refine introduction to the specification's concepts (#6)
This rearranges, and rewrites part of the introduction to Open Job Description
concepts. Notably, the pages on Job Structure, Format Strings, and Overview
have been merged in to a page on How Jobs Are Constructed.

With this change we have a concept introduction that is broken up in to two parts:
1. Job definition and injest; and
2. Running jobs.

A future addition will add a guided narrative introduction that guides the reader
through construction of a job template; starting with a simple one, and then adding
to that through the narrative. Concepts for jobs, job templates, and running jobs
will be introduced incrementally through this narrative form.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>